pyproject.toml
[tool.poetry]
name = "awsiotclient"
version = "0.2.2"
description = "The AWS IoT Device Client provides device-side functionality for AWS IoT services such as jobs, device shadow, and simple pubsub."
authors = ["WHILL Inc."]
license = "Apache License 2.0"
repository = "https://github.com/whill-labs/aws-iot-device-client-python"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7.1"
awsiotsdk = "^1.5.11"

[tool.poetry.group.dev.dependencies]
boto3 = "^1.26.155"
black = "^23.3.0"
flake8 = {version = "^6.0.0", python = ">=3.8.1"}
flake8-pyproject = "^1.2.3"
isort = {version = "^5.12.0", python = ">=3.8"}
mypy = "^1.3.0"
pytest = "^7.0.0"
pytest-cov = {version = "^5.0.0", python = ">=3.8,<4.0.0"}

[tool.pytest.ini_options]
testpaths = ["test"]
# Disable ROS2 (jazzy) pytest plugins that interfere when installed system-wide
addopts = "-p no:launch_testing -p no:launch_ros -p no:ament_copyright -p no:ament_xmllint -p no:ament_pep257 -p no:ament_mypy -p no:ament_lint -p no:ament_flake8"

[tool.isort]
profile = "black"

[tool.flake8]
ignore = "E501"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
